+++ ChangeLog.md
@@ -9,6 +9,8 @@
     - Introduce Forms.FieldType.C.TRANSPARENT so that NONE and TRANSPARENT are both possible and different semantics, thus bumping minor candidate version number
     - Extend crossCheckFn instanciation paramater to also accept an array of functions
     - Check functions now receive the Checker instance as an option member
+    - Define new Checker.panel() and Panel.objectData() methods
+    - Fix the crossChecks behaviours
 
 ### 1.3.0